+++ os/src/task/task.rs
@@ -2,6 +2,8 @@
 
 use super::TaskContext;
 use crate::config::MAX_SYSCALL_ID;
+//extern crate alloc;
+//use hashbrown::HashMap;
 /// The task control block (TCB) of a task.
 #[derive(Copy, Clone)]
 
@@ -11,10 +13,11 @@ pub struct TaskControlBlock {
     /// The task context
     pub task_cx: TaskContext,
     /// The task syscall count
-    pub syscall_count: [usize; MAX_SYSCALL_ID]
-
-
+    pub syscall_count:[u8;MAX_SYSCALL_ID]
 }
+
+
+
 
 /// The status of a task
 #[derive(Copy, Clone, PartialEq)]
